+++ src/main.c
@@ -1,16 +1,31 @@
 #include "minirt.h"
 #include <stdio.h>
 
+void	print_stuff(t_map *map);
+
 int	main(int argc, char **argv)
 {
 	char	**file;
+	int		fd;
 	t_map	*map;
-	int		i;
 
 	if (check_input(argc, argv))
 		return (EXIT_FAILURE);
-	file = get_file(argv[1]);
+	fd = open(argv[1], 0);
+	if (fd == -1)
+		return (ft_putendl_fd("Error: file: not accessable", 2), 1);
+	file = get_file(fd);
 	map = get_map(file);
+	print_stuff(map);
+	setup_window(map);
+	free_map(map);
+	return (EXIT_SUCCESS);
+}
+
+void	print_stuff(t_map *map)
+{
+	int	i;
+
 	i = -1;
 	while (++i < (int)map->light_count)
 	{
@@ -25,7 +40,4 @@ int	main(int argc, char **argv)
 		printf("Color		%d: %f	%f	%f\n\n", i + 1, map->objects[i].color.x,
 			map->objects[i].color.y, map->objects[i].color.z);
 	}
-	setup_window(map);
-	free_map(map);
-	return (EXIT_SUCCESS);
 }
