+++ src/parser/getter.c
@@ -72,3 +72,20 @@ int	get_err_num(int ecode, char *str, t_count count)
 		ecode = ecode + count.cy * 100;
 	return (ecode);
 }
+
+char	*stra_iteri(char **arr, char *set, int id)
+{
+	size_t			j;
+	static size_t	i[3] = {0, 0, 0};
+
+	j = 0;
+	while (arr[i[id]] != NULL)
+	{
+		if (!ft_strncmp(arr[i[id]] + j, set, ft_strlen(set))
+			&& ft_iswhitespcs(arr[i[id]][j + ft_strlen(set)]))
+			return (arr[(i[id])++]);
+		j = 0;
+		(i[id])++;
+	}
+	return (NULL);
+}
